pyproject.toml
[tool.poetry]
name = "snmp-json"
version = "0.1.0"
description = ""
authors = ["James Hodgkinson <james@terminaloutcomes.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pysnmp-lextudio = "^6.1.2"
pydantic = "^2.7.0"
click = "^8.1.7"
pydantic-settings = "^2.2.1"
loguru = "^0.7.2"


[tool.poetry.group.dev.dependencies]
ruff = "^0.3.7"
pytest = "^8.1.1"
mypy = "^1.9.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
"snmp-json" = "snmp_json.__main__:cli"
